% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concentration.R
\name{get_concentration}
\alias{get_concentration}
\title{Get concentration from absorbance data}
\usage{
get_concentration(
  treatments,
  condition_column,
  measure_column,
  plate_column,
  model,
  data
)
}
\arguments{
\item{treatments}{vector with strings for each condition/treatment in the study}

\item{condition_column}{name of the column where the conditions are listed}

\item{measure_column}{name of the column where the measures, corrected by removing the background, are recorded}

\item{plate_column}{column in data that indicates from which plate each data comes from}

\item{model}{named list of lm models, the names of the list should be the same as the strings in}

\item{data}{name of the dataframe with condition_column and measure_column}
}
\value{
a data frame equal to the one entered (as data), but with an additional column with the concentration
}
\description{
After calculating the model with get_model, it uses the measured values of absorbance (with background subtracted already) to get the concentration
}
