% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_background.R
\name{remove_background}
\alias{remove_background}
\title{Removes background from absorbance data}
\usage{
remove_background(
  background,
  treatments,
  condition_column,
  measure_column,
  plate_column,
  data
)
}
\arguments{
\item{background}{string in the condition column that indicates which measures are the blank}

\item{treatments}{vector with strings for each condition/treatment in the study}

\item{condition_column}{name of the column where the conditions are listed}

\item{measure_column}{name of the column where the measures, corrected by removing the background, are recorded}

\item{plate_column}{column in data that indicates from which plate each data comes from}

\item{data}{name of the dataframe with condition_column and measure_column}
}
\description{
This function allows you to subtract the background (when it was measured) for all other measures
}
