% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmbg_gm_gc.R
\name{rmbg_gm_gc}
\alias{rmbg_gm_gc}
\title{Remove background, get model, get concentration}
\usage{
rmbg_gm_gc(
  data,
  background,
  treatments,
  condition_column,
  measure_column,
  plate_column
)
}
\arguments{
\item{data}{name of the dataframe with condition_column and measure_column}

\item{background}{string in the condition column that indicates which measures are the blank}

\item{treatments}{vector with strings for each condition/treatment in the study}

\item{condition_column}{name of the column where the conditions are listed}

\item{measure_column}{name of the column where the measures, corrected by removing the background, are recorded}

\item{plate_column}{column in data that indicates from which plate each data comes from}
}
\description{
This function removes background absorbance from absorbance measured in each treatment (exported as first object of the list),
then it calculates the standard curve from the values in lines for which the condition is a number (considered the known concentration),
the curve is exported as the second object of the list.
Finally, it uses the measured values of absorbance (with background subtracted already)
to get the concentration and saves it to a data frame (only the for the treatments),
this data frame is the third object of the list
}
